Hantek source: Split files further
